+++ core-java-io/.gitignore
@@ -1,26 +1,7 @@
-*.class
-
 0.*
-
-#folders#
-/target
-/neoDb*
-/data
-/src/main/webapp/WEB-INF/classes
-*/META-INF/*
-.resourceCache
-
-# Packaged files #
-*.jar
-*.war
-*.ear
 
 # Files generated by integration tests
 *.txt
 backup-pom.xml
 /bin/
 /temp
-
-#IntelliJ specific
-.idea/
-*.iml
