+++ core-java-io/src/test/resources/targetFile.tmp
@@ -1,0 +1,1 @@
+Hello World
