--- core-java/src/main/resources/targetFile.tmp
@@ -1,2 +1,0 @@
-line 1
-a second line
