--- core-java/src/test/resources/copyTest/dest/readme.txt
@@ -1,2 +1,0 @@
-files will be copied here and then deleted
-remove `file.delete()` to see the files here
