--- core-java/src/test/resources/copyTest/src/test_apache.txt
@@ -1,1 +1,0 @@
-apache
