--- core-java/src/test/resources/copyTest/src/test_channel.txt
@@ -1,1 +1,0 @@
-channel
