--- core-java/src/test/resources/copyTest/src/test_files.txt
@@ -1,1 +1,0 @@
-files
