--- core-java/src/test/resources/copyTest/src/test_stream.txt
@@ -1,1 +1,0 @@
-stream
