--- core-java/src/test/resources/fileToRead.txt
@@ -1,1 +1,0 @@
-This is a content of the file
