--- core-java/src/test/resources/testFolder/sample_file_1.in
@@ -1,1 +1,0 @@
-Hello world
